x/anpanchain/types/messages_people_test.go
package types

import (
	"testing"

	"github.com/anpan2val/anpan-chain/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgCreatePeople_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreatePeople
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreatePeople{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreatePeople{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgUpdatePeople_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgUpdatePeople
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgUpdatePeople{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgUpdatePeople{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgDeletePeople_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgDeletePeople
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgDeletePeople{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgDeletePeople{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
